+++ examples/dagster_examples/schedules.py
@@ -55,7 +55,10 @@ def backfill_should_execute(context, partition_set_def, schedule_name):
     is_remaining_partitions = bool(available_partitions.difference(satisfied_partitions))
     if not is_remaining_partitions:
         try:
-            context.instance.stop_schedule(context.repository, schedule_name)
+            # Instance methods for the scheduler require both the repository name and
+            # the schedule name. Here, we hardcode the repository name that this
+            # schedule belongs to.
+            context.instance.stop_schedule('internal-dagit-repository', schedule_name)
         except OSError:
             pass