+++ python_modules/dagster-graphql/dagster_graphql/implementation/execution/scheduled_execution.py
@@ -45,8 +45,8 @@ def start_scheduled_execution(graphene_info, schedule_name):
         # We first load the repository and schedule definition to create
         # and store a ScheduleTick.
         # If this fails, this error should be sent to the file based scheduler logs.
-        repository = graphene_info.context.get_repository_definition()
-        repository_name = repository.name
+        external_repository = graphene_info.context.get_external_repository()
+        repository_name = external_repository.name
         schedule_def = get_dagster_schedule_def(graphene_info, schedule_name)
         cron_schedule = "Unknown" if not schedule_def else schedule_def.cron_schedule
         tick = graphene_info.context.instance.create_schedule_tick(
@@ -60,7 +60,7 @@ def start_scheduled_execution(graphene_info, schedule_name):
         )
 
         # Run should_execute and halt if it returns False
-        schedule_context = ScheduleExecutionContext(graphene_info.context.instance, repository)
+        schedule_context = ScheduleExecutionContext(graphene_info.context.instance)
         with user_code_error_boundary(
             ScheduleExecutionError,
             lambda: 'Error occurred during the execution should_execute for schedule '