+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_schedules.py
@@ -69,9 +69,7 @@ def get_schedule_or_error(graphene_info, schedule_name):
 
 
 def execution_params_for_schedule(graphene_info, schedule_def, pipeline_def):
-    schedule_context = ScheduleExecutionContext(
-        graphene_info.context.instance, graphene_info.context.get_repository_definition()
-    )
+    schedule_context = ScheduleExecutionContext(graphene_info.context.instance)
 
     # Get environment_dict
     with user_code_error_boundary(